LUN-399: moved AudioRecordOverlay to route/components
